Create Flatten Linked List.py
Flatten a Linked List
Solved Medium Success : 89% Helpful
You're given the head of a linked list of size N , where each node represents a sub-linked list and contains two pointers:
A next pointer to the next node,
A down pointer to a linked list where this node is the head.
Each sub-linked list is sorted in ascending order. Your task is to return the head of the flattened linked list, where all nodes appear on a single level in
sorted order.
Example 1:
Input:
Given a Linked list:
Output:
Explanation:
The output is a single-level linked list where all the nodes are sorted in ascending order. The nodes should appear in the following sequence:
1 -> 2 -> 3 -> 4 -> 5 -> 6 -> 7 -> 8 -> 9 -> 10 -> 14 -> 15